feat: move find_free_port to _helpers.py and add helpful logger message for Windows users
